Workout Page Dynamic Image Loading
Whenever you select a specific exercise, the workout page will now display an image appropriate to the workout.  It is rather rough, limiting image size to 300x300 so it is cropping many of the images.  However, that just requires a style update to fix.

I've also changed a bit of the backend so that the images are being served by our Django server and stored there.  This makes it very easy for us to change update the workout page with more appropriate images.